backend/requirements.txt
Flask==2.2.5
Flask-JWT-Extended==4.4.4
Flask-Cors==4.0.1
SQLAlchemy==2.0.20
psycopg2-binary==2.9.7
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
